+++ CHANGES.md
@@ -1,3 +1,8 @@
+## GRNOC NetSage Deidentifier 0.0.2 -- Mon July 25 2016
+
+### Features:
+ * ISSUE=635 PROJ=160 Added flow stitching functionality. 
+
 ## GRNOC NetSage Deidentifier 0.0.1 -- Fri April 1 2016
 
 ### Features: